Backend/requirements.txt
fastapi~=0.111.0
uvicorn
sqlalchemy~=2.0.30
python-dotenv~=1.0.1
mysqlclient
pydantic~=2.7.1